Deep Learning: Add documentation for helper function and refactor code.
JIRA: MADLIB-1306

This commit adds user and online documentation (without examples) along
with addressing other comments from the code review for PR #356.

As part of code refactoring, this commit also makes deep learning a
new module instead of creating new .py_in and .sql_in files in existing
convex module. This will make it easier to control which databases we
install it in. For instance, DL will not be supported on GPDB-4.3, so we
can easily avoid installing that whole module on GPDB 4.3. This will
also help in better maintenance by having future DL functions in this
module rather than in convex.

Closes #356