Create a new model for deep_learning instead of using convex
This commit makes DL a new module instead of creating new py and sql files in
convex module. This will make it easier to control which databases we
install it in. For instance, DL will not be supported on GPDB-4.3, so we
can easily avoid installing that whole module on GPDB 4.3. This will
also help in better maintenance by having future DL functions in this
module rather than in convex.