Deep Learning: Disable deep learning for gpdb 4.3
We won't support deep learning on gpdb 4.3 because
1. Deep learning requires the type 'JSON' which isn't supported in gpdb
4.3.
2. Our deep learning module depends on tensorflow to be pre installed.
GPDB 4.3 ships with python 2.6.9 which doesn't support tensorflow.

madpack currently looks for Modules.yml first in
"src/ports/portid/dbver/config" and then in src/config. This commit
copies the custom Modules.yml in the src/ports/portid/dbver/config/ dir.
We use a custom Modules.yml file for gpdb 4.3 which is just a copy of
the existing Modules.yml file.

A future commit will add support for the 'deep_learning' module which
will not be included in this yml.

Closes #358

Co-authored-by: Nikhil Kak <nkak@pivotal.io>