DL: Remove reshaping from training and validation data
JIRA: MADLIB-1304

We now assume that the user has the correct input shape to match the
model architecture. We will not reshape either training or validation
data for preparing data for keras.fit/keras.eval

We now run keras.eval on segments instead of master. This is done to
make use of the MPP architecture.  We now assume that the validation
data is not minibatched i.e. one image per row.  We run keras.eval on
each row of the validation data and then aggregate the results using
madlib.array_avg.

Additionally
1. Add import keras although it's not directly used.  For ex if the user
passes in the optimizer as keras.optimizers.SGD instead of just SGD,
then without this import our keras python files won't find the SGD
module.
2. Moved the hardcoded value of gpus_per_host to the
`get_device_name_for_keras` function. This will eventually be refactored
in a future PR.
3. Add numeric array check to dependent var for validation table.

Closes #367
Co-authored-by: Jingyi Mei <jmei@pivotal.io>
Co-authored-by: Orhan Kislal <okislal@apache.org>