Deep Learning: Use compile_and_set_weights() in predict
Commit SHA 137ba49 changed the way we process compile_params, and
although it was used in fit, it wasn't used in the predict function.
This commit makes necessary changes to process compile_params using this
new function. We also now use KerasWeightsSerializer.get_model_shapes to
get the model shape in predict function.