Fix bugs, add test and newline
This commit adds a newline at the end of the file, and fixes 3 bugs:

1. Change ' to " on line 191
2. Rename keras_out in 1st call to keras_fit to keras_out_saved, and
   use that for predict.  keras_out was being overwritten later,
   causing inaccurate predictions.
3. Change $$ to $TRAP$ on lines 200,208

It also adds an additional test that the validation accuracy has
reached 100% after training.  This should be easy to reach after
3 iterations, since there are only 2 training examples.  If it
doesn't reach it, then something's broken.  (This should make
it easier to diagnose the problem if the predict results are
inaccurate in the future.)