WCC: Duplicate edges to improve performance on Greenplum
JIRA: MADLIB-1320

Prior to this update, we used the edge table in join clauses with both
src and dest coulmns.

This commit updates the query on the message table such that the JOIN
always operates on the column that the table is distributed on. To this
end, we duplicate the edge table to distribute it on the dest column.
This seperation of accesses reduces the redistribute motion on gpdb.

In addition, we split the union all query to ensure that the working set
is limited with very large graphs.

Closes #364

Co-authored-by: Orhan Kislal <okislal@apache.org>