DL: Add new columns in fit output summary table.
JIRA: MADLIB-1319

Columns `dependent_vartype` and `normalizing_const` are useful to have
in the model summary table since they can be used in predict.
1) For instance, `dependent_vartype` will be useful in
https://issues.apache.org/jira/browse/MADLIB-1315 while creating the
prediction output column. In case of response, the output column should
be of the same type as the dependent variable in train data.
2) While predicting, `normalizing_const` will be useful to normalize the
test data using the same normalizing constant that was used to normalize
the train data.

Closes #365

Co-authored-by: Nandish Jayaram <njayaram@apache.org>