DL: Remove evaluate function
JIRA: MADLIB-1304

The `evaluate` UDF keeps all the validation table data in master's
memory and then calls keras.eval on it. We can instead run keras.eval on
all the segment hosts so that we don't have to keep all data in in one
node's memory. We already have a UDF named `evaluate` that is
implemented with this logic which will be renamed to `evaluate` in a
future PR.

This commit removes the evaluate function from the sql as
well as the python file.

Closes #367

Co-authored-by: Jingyi Mei <jmei@pivotal.io>