DL: Add GPU support for evaluate
In our experiments on a multi host gpdb cluster with places10 (subset of
places365) dataset, we noticed that using CPU and no SD caused the fit
query to fail during the validate code with the error 'Dst tensor is not
initialized'. Adding GPU and caching the compiled model fixed this issue.

1. Add GPU support to the evaluate code called by the fit function.
2. The evaluate code previously compiled the model for each row in the validation table.
This was unnecessary and takes more resources in terms of time and memory. This commit avoids this by using SD to cache the compiled model.
3. Also refactored the code by creating a function to get the rows per segment