DL: Replace use_gpu flag with gpus_per_host integer
JIRA: MADLIB-1308

Previously, gpus_per_host were hard coded to 4. This commit removes this
hard coding and takes in this value from the user.

We also tried to use the tensorflow function `list_local_devices` to get
the count of gpus per host. This did give us the count but would hang
forever on some segments. So we decided to not use this function.

We now cache the CUDA_VISIBLE_DEVICES env variable (which is set to
-1 for master) and then reset it at the end of fit function.

Closes #379

Co-authored-by: Nikhil Kak <nkak@pivotal.io>