DL: Update madlib_keras_fit code to weight by images instead of buffers.
JIRA: MADLIB-1310
Prior to this commit, while merging the weights, accuracy and loss, we
assume that all the buffers are of the same size. This may not be always
the case.  Ideally we should use the total no of images per segment to
average the weights, accuracy and loss.
This commit, updates the transition, merge and final function for
`madlib_keras_fit()` to use the number of images on a segment instead of
the buffer count for averaging weights, accuracy and loss.

Additionally, the function `deserialize_weights()` when deserializing
loss and accuracy from state, type-casted it to an int, which resulted
in computing lower loss and accuracy while training. This commit also
addresses this issue.

Closes #378
Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>