DL: Remove loss and metric calculations from fit transition function
JIRA: MADLIB-1338

This commit removes the loss and metric calculation from the fit transition
function to improve performance. For now, we output in the metric column
as 0.0 in the summary table if metric is not passed in, and we will make
a PR later to distinguish 0 and None value later.

In addition, it updates some variable names to reflect supported
metrics, which previously were only 'accuracy'.

Finally, we added dev-check tests for metrics other than accuracy.

Closes #393
Co-authored-by: Jingyi Mei <jmei@pivotal.io>
Co-authored-by: Nandish Jayaram <njayaram@apache.org>
Co-aurthored-by: Ekta Khanna <ekhanna@pivotal.io>