DL: Update content of the model summary table
JIRA: MADLIB-1349, MADLIB-1338
Add a new column named `metrics_type` that captures the metric name
(if) specified in the compile params. This commit also fixes the
model_size reported, it was reporting the wrong size.

There were a few comments in MADLIB-1338 after the corresponding PR was
merged to master. Some of those comments are addressed in this PR since
they are related code, and minor changes.
1. Ensure compile_params cannot be NULL.
2. Update some info messages.
3. Use the @MinWarning decorator to suppress notices.
4. Ensure training_metrics_final and validation_metrics_final are NULL
if the compile params did not have any metric specified.

Closes #397
Co-authored-by: Jingyi Mei <jmei@pivotal.io>