DL: fix default_buffer_size calculation
JIRA: MADLIB-1340

In buffer size calculator, we assumed the array to be packed is one dimention, and use the array_upper(x, 1) to get the length
of the array. In DL input data preprocessor, we passed the first element
in array_ndims as the length, which is not right because the
array can be multi-dimentional. This caused
default_buffer_size_calculator returns a bigger buffer size than it
supposed to have and rows get too big and crashed database.

Instead, we should use
product of all the elements from array_ndims to represent the actual length of
the array. For example, if array_ndims returns [32,32,3], we should pass
32*32*3 instead of 32. This commit fixes this issue by passing the
right length to defualt buffer calculator.

Co-authored-by: Nikhil Kak <nkak@pivotal.io>