DL: fix default_buffer_size calculation in input_data_preprocessor
JIRA: MADLIB-1340
In buffer size calculator, we assumed the array to be packed is one
dimensional, and used the array_upper(x, 1) to get the length
of the array. In DL input data preprocessor, we passed the first element
in array_ndims as the length, which is not right because the
array can be multi-dimensional. This resulted in
default_buffer_size_calculator returning a bigger buffer size than it
was supposed leading to the 1GB limit error.

Instead, we should use product of all the elements from array_ndims to
represent the actual length of the array. For example, if array_ndims
returns [32,32,3], we should pass 32*32*3 instead of 32. This commit
fixes this issue by passing the right length to defualt buffer calculator.

Closes #401
Co-authored-by: Nikhil Kak <nkak@pivotal.io>