SVM: Fix class weights when specified as a mapping
JIRA: MADLIB-1346
Providing a mapping for the class_weight param was resulting in an error
since the dependent var's levels were not quoted when used in an
internal query. This PR now quotes the values.

Additionally, this PR adds the following new error checks for the
class_weight param:
1. Permitted values are `balanced`, a mapping of the form `{key:value}`,
empty string.
2. Keys in a mapping must be legitimate class labels in the source
table.
3. Values in a mapping must be numbers.
4. There has to be at least one key-value and at most 2 key-value pairs.

This PR still does not address the issue of having special characters in
class levels used in a mapping. MADLIB-1354 was created to track that
issue.

Closes #403