Fix tensorflow crashes
This adds `std::*` to the list of symbols defined in madlib
that should be hidden from external libraries by the linker.

This fixes several intermittent crashes that were happening in the
deep_learning module in different forms on different platforms, due to
tensorflow calling madlib's instantiation of STL template functions,
which circumvents the hiding of madlib's global new and delete symbols
(intended to override the libstdc++ version of these symbols for all
of `libmadlib.so`, but not for anything outside of it).

The main crash we looked at was due to the use of `std::set<string>` in
`madlib/src/modules/linalg/metric.cpp`.  Tensorflow is a python library,
but it loads Google's C++ protobuf library `_message.so`.  The crash was
happening because `_message.so` also uses `std::set<string>` for some
things.  We were able to fix it by hiding all madlib symbols associated
with the `std::set<string` class, but this simply resulted in a
different crash happening instead.

In the long term, we should move to hiding all symbols by default
and explicitly make exceptions for madlib API functions.  But this would
be a bigger change and require more extensive testing to make sure it
doesn't cause any obscure issues on any platform.  So for now we are just
hiding those in namespace std.

Closes #404

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>