DL: Fix predict with 'NULL' string class values
JIRA: MADLIB-1357
Fix handling of 'NULL' string for class values in predict. 'NULL' string
as a class value was getting reported the same way as a Postgres NULL
class value (i.e., empty). This commit double-quotes a 'NULL' string
class value in the relevant query.
NOTE: In predict, if we use 'prob' as the pred_type, we create a column
for each distinct class level. The column name for Postgres NULL class
level is 'prob_NULL', and the column name for a 'NULL' string class
value will be 'prob_"NULL'.

Closes #408