Association Rules: Improve performance
JIRA: MADLIB-1327
Assocation rules was slow due to a blow up in number of candidate
itemsets before checking for their support and graduating them to
frequent itemsets. This PR changes ensures the candidate itemsets are
ordered, and only a few consecutive itemsets must be considered to merge
and create a new candidate itemset for checking support. This also
reduces the result of a join query significantly (previously, the result
of the join were all potential candidates, which were redundant and
which may have resulted in itemsets that were larger than what was
considered in the current iteration).
This commit also adds relevant tests in dev-check.

Closes #423
Co-authored-by: Orhan Kislal <okislal@apache.org>