DL: Split madlib_keras devcheck file
Prior to this commit, the dev-check test file for deep_learning module
was too big and there was no way of running a subset of the tests for
e.g. predict.
This commit separates madlib_keras.sql_in test file into:
1. Setup files: for data creation used in the tests
2. Specific test files: for testing fit, predict, evaluate and transfer
learning

Additionally, madpack.py is updated to ignore any setup
files(`*.setup.sql_in`) in the test dir and not run it as a test in
dev-check/install-check.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>
Co-authored-by: Orhan Kislal <okislal@apache.org>