KNN: Add distances to the output table
JIRA: MADLIB-1370

This commit adds a new column to the output table to expose the
distances for the nearest neighbors. This column shows up only when the
user actually requests nearest neighbors.

In addition, this commit adds a safeguard for the distances. We get the
absolute value of the distance function output in case some floating
point operations cause the distance function to return a negative value.

Closes #427

Co-authored-by: Frank McQuillan <fmcquillan@pivotal.io>