MADLIB-1351 : Added stopping criteria on perplexity to LDA
Prior to this commit, in LDA there are no stopping criteria. It runs for
all the provided iterations. This commit calculates the perplexity on
each iteration and when the difference between the last two perplexity
values is less than the perplexity_tol, it stops the iteration.

These are the two new parameters added to the function:

```
evaluate_every  INTEGER,
perplexity_tol  DOUBLE PRECISION
```

Also, there is a change to the model output table. The following new
columns are added:

1. perplexity(DOUBLE PRECISION[]): is an array of perplexity values as
per the 'evaluate_every' parameter.
2. perplexity_iters(INTEGER[]): is an Array indicating the iterations
for which perplexity is calculated