DL: Remove model_arch_table from output of load_model_selection_table()
JIRA: MADLIB-1381

The utility function `load_model_selection_table()` assumes that there
is only 1 model architecture table (passed in as a parameter).
Therefore, it should not be added in the output table.
This commit removes `model_arch_table` column from the output table of
`load_model_selection_table()` and creates a summary table with this
column instead.
Also, added tests and split the unit test file.

Co-authored-by: Ekta Khanna  <ekhanna@pivotal.io>