DL: Improve performance for madlib_keras_predict()
JIRA: MADLIB-1345

Passing huge model weights as a param to `internal_keras_predict()` for
each table row slowed the performance of overall
`madlib_keras_predict()`. With this commit, the model weights will only
be passed in as a param for the very first row(min(ctid)) fetched on
each segment and NULL for the rest. With this change, we see `~4x`
performance boost in the execution time of the existing
`madlib_kerase_predict()`

Co-authored-by: Nikhil Kak <nkak@pivotal.io>