DL: Update evaluate and predict for multi model outputs
JIRA: MADLIB-1387

This commit updates the evaluate and predict functions to support individual
models from fit_multiple outputs.

Both of the functions work by calling the vanilla versions. They create a view
from the existing summary and info tables and then pass it as the summary
table.

This commit includes following changes as well:
- Remove image_count from multi model agg state. This value was not used since
we take the count from the prev state.
- Add a final function for fit multiple to remove the image_count.
- Rename model_data to model_weights to be more consistent.
- Removed json.dumps() call since it was just adding extra quotes and was
unnecessary.
- Add validation tests for evaluate and predict.