DL: Avoid having lingering processes from previous query
Creating a TEMP table is a session level command that ensures the
temporary table created is visibile throughout the session it gets
created in. This is acheived by keeping the slice process, which created
the temp table, alive for the entire session.
In DL, the same process initializes the GPU memory, which is not
released until the process goes away. Since the process is alive for
the entire session, this can cause the subsequent query to fail as it
tries to initialize GPU memory which is already allocated to the
previous queries process.
This commit fixes this issue by not creating any TEMP tables.

Co-authored-by: Nikhil Kak <nkak@pivotal.io>