DL: Add param is_minibatched to validate_input_shape
Prior to this commit, the function for validating input shapes (common
to evaluate, fit and predict) was checking if the input data was
minibatched or not, by validating the _shape column in it. This was
givin a warning for predict, as _shape column doesn't exist.
This commit adds a new param `is_minibatched` to the function so that
the caller validates if the data is minibatched or not and passes in the
param accordingly.

Co-authored-by: Nikhil Kak <nkak@pivotal.io>