Build: Add support for pg 12
JIRA: MADLIB-1391

This commit adds the necessary files for PG12 support and makes a few changes
in the backend.

The FunctionCallInfoData struct from postgres has been changed to
FunctionCallInfoBaseData. This new structure replaces the "value" variable with
"value" and "isNull". We made changes in the backend to handle this new
struct.
Also, `get_float8_infinity()` was replaced with using FLT_MAX to
accommodate the changes for the PG12 commit postgres/postgres@6bf0bc8.

Co-authored-by: Nikhil Kak <nkak@pivotal.io>