DL: Add asymmetric cluster support for fit and evaluate
JIRA: MADLIB-1393

This commit updates the fit and evaluate transition functions to support
asymmetric clusters.

The gpus_per_host variable is changed to use_gpus. This boolean variable is
used to indicate if CPUs or GPUs are used for keras operations. The gpu
configuration to use is read from the preprocessor summary table.

Closes #462

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>