DL: Update tests calling trap_error
Prior to this commit, due to the missing `SELECT`, call to
`trap_error()` would always return 1, irrespective of whether
the execution of the query would error out or not. This commit
updates such tests to pass in an executable SQL query.
This commit also fixes a bug in predict that was missed due to
the missing `SELECT` in the query param passed to trap_error().