Graph: Fix output table schema parsing bug
JIRA: MADLIB-1411

HITS and Pagerank functions failed if the output table had a schema
attached to it. This error was caused by the final alter table operation,
since postgres does not allow users to alter a tables schema.

This commit changes this operation to create table and adds tests to
every graph function to make sure we don't have similar bugs in the future.