DL: Fix disk issue by using truncate guc
JIRA: MADLIB-1406

While testing places10 with fit multiple (gpdb5 and gpdb6, 10 iterations and 20 msts), we ran out of disk space although we had at least 1.5T left at the beginning of the query.

The main contributor to the disk bloat is the update statement that we run at the end of each hop

```
UPDATE {self.model_output_table}
SET {self.model_weights_col} = {self.weights_to_update_tbl}.{self.model_weights_col}
FROM {self.weights_to_update_tbl}
WHERE {self.model_output_table}.{self.mst_key_col} = {self.weights_to_update_tbl}.{self.mst_key_col}
```

In postgres/gpdb, every update command is really two commands i.e. insert and then delete. Because of this, the actual space is not freed and only gets freed when vacuum is run consistently or vacuum full is run at the end. We verified this by printing the {self.model_output_table} size for each mst_key and it kept on growing with each update statement.

Also the disk space for other intermediate tables that get created in the run_training function never gets cleared even though we drop these tables inside the said function. This is because drop/truncate does not release disk space inside a pl function since it's in a sub transaction (this is so that it can rollback). It only releases the space once the pl function has completed execution.

The only way to make this work was to add a truncate statement and change the gpdb code to do a truncate inside a sub transaction. gpdb 6.5 introduced a guc https://github.com/greenplum-db/gpdb/commit/b4692794a0abd8d8051c23019d26e22f7f3d0aa5 which when turned 'on' allows for truncating the disk space inside a sub transaction.

Note that this guc is only available in gpdb 6.5 and up.

run_training workflow (this function is called per hop)
1. join schedule table with mst_weights table to do the hop
2. Call the uda and copy the output to an intermediate table
3. Update the model table with the results of the previous step
4. set the truncate guc to on
5. Create temp table from model table
6. truncate the model table to release disk space
7. rename temp table to model table so that it can be reused for the next hop

Warm Start:
For warm start, we can't keep calling truncate on the user passed output table
because then we won't be able to roll it back in case of a failure. So for warm
start, we create a copy of the output table passed by the user and then operate
on the copied table. At the end, we drop the original output table and rename
the copied table to the original table name.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>