Remove tests for no GPU hardware
This commit removes 2 problematic tests that we decided weren't worth including in the DL dev-check tests.

One of these tests was checking to make sure that madlib_keras_fit() fails when you pass use_gpus=True. And the other does the same for madlib_keras_predict().  This wasn't a problem originally because we usually run dev-check only on Jenkins or other docker images where there are no GPU's. But as we continue to improve our pre-release testing process, we'd like to have dev-check be able to pass on GPU-enabled test platforms as well. So these tests are no longer appropriate.