Graph: Fix column type concat bug
JIRA: MADLIB-1408

The apsp get path function fails if the the input columns are of bigint type.
The issue stems from the concatenate operation we use to build the array that
represents the path. Adding an integer to an integer array works as expected
but adding a bigint to an int array (or vice versa) doesn't.

This commit fixes the issue by detecting the appropriate type and adding the
proper cast operations.

Closes #483