DL: Set plan_cache_mode when passing weights for GPDB6
JIRA: MADLIB-1414

For versions >=GPDB6, previously, for queries called with weights value
passed in as a param, the query plan for it would create custom plans
with weights embedded in the plan itself.  This meant that the query
plan size would also include the size of these weights, bloating it up
to hit the 1GB limit when dispatching the query plan to segments,
leading to OOM for large weights.

In GPDB, for PREPARE plans, there is a GUC `plan_cache_mode` that
controls choosing generic plan(not constant folding the passed in
params)/custom plans(constant folding the passed in params) for
execution. Therefore, for GPDB6 setting this GUC to `force_generic_plan`
for using generic plans when passing in weights as params to queries.