DL: Don't include weights as part of state except for the last row.
JIRA: MADLIB-1418

Previously the model state would consist of the image count and the serialized
model weights but there was no need to include the serialized model weights as
part of the state except for the last row for that segment. We were only using
that state to get the image count and ignored the model weights, so serializing
the model weights for every row was a waste to time and resources.

This commit changes that pattern so that for all rows except the last row, the
state would be just the image count only.