Utilities: Improve rename table logic
JIRA: MADLIB-1406

Consider the example
"rename table foo to bar.foo2"

Previously rename_table would fail if the table 'foo2' already existed on the
search_path. This is because we first renamed the table foo to foo2 and then
set the schema to bar.

The new logic fixes this by renaming foo to a unique string, setting the schema
to bar and then renaming the unique string to bar

Also moved the function to utilities since the new logic required the use of
unique_string. Also it seemed better suited to utilities.py instead of
validate_args.py

Co-authored-by: Domino Valdano <dvaldano@pivotal.io>
Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>