Build: Remove function dependencies from changelist
Currently the madpack checks if there are any dependencies on madlib functions
(and aggregates). This has 2 steps
1. Check if there is a dependency on a madlib function
2. Check if the function is included in the changelist
If both conditions are satisfied, madpack stops the upgrade and gives a proper
message to the user including the dependency chain.

This commit removes the second step (changelist based check) and stop the upgrade
if there are any dependencies on madlib functions. It will ensure that our upgrades
are less error prone and the user will still have the dependency chain.