Add conditional support for C++11 compilers.
This allows madlib to be compiled on OSX with libc++ and
other C++11 compilers.

On OSX, this gets rid of the need to use either an old version
of MacOS (Sierra or earlier), download an old version of
XCode (9.x or earlier), or compile with gcc (the last of which
is unacceptable for a release version, as it means it will not
run on MacOS without the user manually installing 3rd party
libraries).

- CXX11 CMake flag added, which enables support for the C++11
  ABI.

- Code has been updated to meet C++11 standards, but for now,
  changes are hidden behind #if macros, and only become active
  when -DCXX11 is passed.

NOTE:  Old versions of Boost (< 1.65) are incompatible with the
  C++ ABI, including Boost 1.61 which is auto-downloaded by
  madlib if none is found.  Soon we should probably
  download a newer version of Boost if it's not detected and
  CXX11 is enabled.  For now, the user must run "brew install
  boost" if they need CXX11 support.