DL: Add support for grid search and random search
JIRA: MADLIB-1439

The load_model_selection_table function requires the user to manually
specify the grid of compile and fit params. Hence, we implement a
function called generate_model_selection_configs (in the same module) to
perform grid/random search.

The user would declare the compile and fit params grid separately as
strings enveloping python dictionaries, and the name of the search
algorithm (along with any corresponding arguments), and the output
format of the new function would be the same as the previous one for
better integrability with other MADlib functions related to model
training etc.

This commit includes the implementation, unit tests (in python and
SQL) and documentation for the newly created function.