DL: [AutoML] Add support for 'diagonal' Hyperband optimized for MPP
JIRA: MADLIB-1447, MADLIB-1448, MADLIB-1449

We integrate AutoML capabilities in Apache MADlib by introducing a
function called madlib_keras_automl, which bridges the worlds of setting
and running model selection together, and helps automate and accelerate
the model selection and training processes end-to-end. The user can
declaratively specify the names of their train/val datasets, mst and
output tables, model architecture and param grid details, the chosen
method name and associated params, and various training details as
desired, and our API handles the scheduling and execution components
with the algorithm workload info displayed to the user.

In the case of MPP databases such as Greenplum, we further accelerate
this algorithm by simultaneously evaluating multiple rounds of the
algorithm located along a 'diagonal', to keep machines busy and take
advantage of the large distributed storage and compute power offered by
Greenplum.