DL: Generate Hyperband schedule for AutoML
JIRA: MADLIB-1445

As part of the new AutoML module for Model Selection with Deep
Learning, we implement a utility function called hyperband_schedule
to generate a viewable schedule for hyperband execution.

This is aimed at users interested in using the hyperband
algorithm and wanting to explore a variety of schedules with
particular R, eta, or skip_last values, to make an informed
decision on the schedule they prefer to use for the actual
execution of the algorithm (which will be supported in an AutoML
PR coming soon).