DL: Improve the predict output with prob (#514)
* DL: Improve the predict output with prob

JIRA: MADLIB-1451

The prob output for predict created a bunch of columns for different
class values. This commit pivots this format so the table signature
is the same but the number of rows is increased thanks to a separate
row for each class value.

In addition, this commit adds the option to filter the output by
a min probability or top n results as well as a ranking column for
non-response options.