Convert EXECUTE to PERFORM
EXECUTE is supposed to act on a string, PERFORM on a query.
I think the EXECUTE was only working because the call to
assert() returns NULL, which I guess is treated as ''
(so it runs SELECT '').  I'm not sure it matters in this
case, but this way is more straightforward.